backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic[email]
passlib
python-jose[cryptography]
python-dotenv
